+++ sources/AsteriskAndAmpersand.cpp
@@ -1,4 +1,5 @@
 // AsteriskAndAmpersand.cpp by Ulrich Eisenecker, April 20 2021
+// AsteriskAndAmpersand.cpp by Ulrich Eisenecker, April 20 2021
 
 int main()
 {
@@ -6,9 +7,4 @@ int main()
    
    double d { 99.0 };
    double & referenceToDouble { d }, aDouble { 100.0 };
-}
-
-void test()
-{
-
 }