+++ CHANGELOG.md
@@ -15,3 +15,13 @@ Updated files:
 ### Modified:
 
 ### Removed:
+## 2025-10-07 14:21:54 UTC
+
+### Added:
+- sources/ReturnVoid - Copy.cpp
+
+### Modified:
+- sources/AsteriskAndAmpersand.cpp
+
+### Removed:
+-